Backport of #2721: Roll forward master tablet deletion on fault during DeleteTable
Summary:
During `DeleteTable`, it is possible that we replicate the table state, but we fail or lose leadership before marking the tablets as deleted. This results in the next master leader loading up the tablets as running, despite their table being deleted.

Test Plan:
Jenkins: auto rebase: no
New unit test. Verified that this test fails without this fix.

Reviewers: mikhail, bogdan

Reviewed By: bogdan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8170