[#4089] Use Raft config as a source for master addresses at master
Summary:
Due to issues in YW after the full move, we could end up with empty master_addresses in server.conf for master.
But we could workaround it by using addresses from Raft config for this master.

Master uses transactions for DDL in YSQL, which requires the client for communication with the transaction coordinator.

Test Plan: ybd --gtest_filter PgMiniTest.MoveMaster

Reviewers: bogdan

Reviewed By: bogdan

Subscribers: kannan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8323