[#3342][YW] Add additional option to specify keyspace when creating backup or scheduled backup.
Summary:
Add additional dropdown for table keyspace. This value is ignored if the user selects one
or more tables, but respected if the user specifies Full Universe backup

Test Plan:
Connect to portal and try creating a backup with the keyspace specified. Confirm that the
backup API call contains "keyspace" as a key. Confirm that user must specify either one or more
tables or select a keyspace in order to create the backup. Try creating a Full Universe backup and a
Full Universe backup with the keyspace specified. Check that any scheduled backup with the keyspace
selected shows up on the /schedules page.

Reviewers: arnav, sanketh, ram

Reviewed By: ram

Subscribers: ui

Differential Revision: https://phabricator.dev.yugabyte.com/D8222