#4309: [YCQL] Make peers_v2 part of readable system tables
Summary:
Currently the cassandra-driver-core:3.8.0-yb-2-RC1 driver tries to access system.peers_v2 table, and if it receives an error that the table doesn't exist, it will try the system.peers table. But currently, by default users don't have permissions on the tables that are part of the system keyspace, but we have added a few exceptions for tables that are needed for the driver to work correctly. This diff adds system.peers_v2 to that list of exceptions.

Currently we check permissions on the system keyspace before we check wether the table exists, so by allowing this system table to be readable, we let the code determine that this table doesn't exist and return the expected error to the driver.

Test Plan: Nikhil tested this fix with the new driver

Reviewers: mihnea, sudheer, nikhil

Reviewed By: mihnea

Subscribers: kannan, nikhil, yql, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D8388