#4213: Fix TestTserverHealthChecks test
Summary: Currently test TestTserverHealthChecks is failing because we no longer leave a tablet in the FAILED state and the test expects it so that it can verify that it can correctly count the number of failed tablets. This diff fixes the issue by using the already existing flag `TEST_delay_removing_peer_with_failed_tablet_secs` to keep the tablet in the FAILED state for the specified amount of time so that the check is successful.

Test Plan: ybd release --java-test org.yb.client.TestTserverHealthChecks#testTserverHealthChecks -n 100

Reviewers: sudheer, bogdan

Reviewed By: bogdan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8395