[YW][#4190][#4172] Add new form for Alerting data so it does not depend on general profile data
Summary:
Add new form for Alerting data that does not require the customer profile email login to be
present in the event that it is null or not an email.

Test Plan:
Try creating a user that not an email such as admin and try setting the alert data for
it. Check whether the alerting data was saved correctly by refreshing the page.

Reviewers: bogdan, wesley, ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D8406