[#4277] Use SeekForward for intents
Summary:
This diff fixes an issue with seeking forward on the intents RocksDB in
IntentAwareIterator. Instead of seeking forward, we were always
performing a seek there, which is not efficient. For instance, in the
newly added test, we insert 1M rows and query number of rows, which
results in a full table scan. Without this fix, it took ~5 seconds, and with
the fix, it is ~3 seconds (~ 1.66x faster)

There was also an inefficiency that was hidden by always using Seek instead of SeekForward.
This bug was caused by lazy intent seek: instead of doing
an actual seek we store the seek request and perform the actual seek
when `Valid` is called on this IntentAwareIterator. Consider the
following secnario. Suppose that Key1 < Key2 < Key3 and the following
sequence of actions is performed on the IntentAwareIterator:

1) Seek to Key3 and check iterator validity. So actual seek is done and
   intent iterator points to Key3.
2) Seek to Key1. No real seek is performed on the intents RocksDB
   iterator.
3) Seek forward to Key2.

Since (2) does not move the intents RocksDB iterator and that iterator
already points past Key2 by the time we reach step (3), no seek will be
performed by step (3) and Key2 would not be found.

So most part of this diff fixes the issue above, and then we just
replace seek with seeking forward in `SeekForwardToSuitableIntent`.

Also improved logging in IntentAwareIterator and added small
optimization that avoids extra seek in `FindLatestRecord` when it was
already performed externally.

Test Plan: ybd --gtest_filter PgMiniTest.BigRead

Reviewers: mikhail, bogdan, timur

Reviewed By: timur

Subscribers: kannan, amitanand, neha, ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D8386