YSQL: create WaitUntilIndexPermissionsAtLeast APIs (#4347)
Summary:

Create a client function `WaitUntilIndexPermissionsAtLeast` to wait for
an index to reach or exceed a given permission.  For example, wait for
index permissions to be at least `INDEX_PERM_WRITE_AND_DELETE`.  Have it
return the actual permission when done waiting.  This is to be able to
handle both success and failure cases.  Continuing the previous example,
a failure can cause the permission to be
`INDEX_PERM_DELETE_ONLY_WHILE_REMOVING`.

Create functions (checked below) to bubble the information up to the
pggate layer.

* [x] `YBCPgWaitUntilIndexPermissionsAtLeast`
  * [x] `PgApiImpl::WaitUntilIndexPermissionsAtLeast`
    * [x] `PgSession::WaitUntilIndexPermissionsAtLeast`
      * [x] `YBClient::WaitUntilIndexPermissionsAtLeast`
        * [x] `YBClient::Data::WaitUntilIndexPermissionsAtLeast`
          * [x] `YBClient::Data::GetIndexPermissions`
            * [ ] `YBClient::Data::GetTableSchemaById`

This is part 1 of the effort of bringing index backfill to YSQL.

Close: #4347

Test Plan: jenkins

Reviewers: amitanand, mihnea

Reviewed By: mihnea

Subscribers: amitanand, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D8398