[#4535] Remove applied intent doc hybrid time during compaction
Summary:
To reduce the frequency of read restarts, we keep the original intent's
DocHybridTime in the value part of key/value records in the regular
RocksDB after the intents have been applied. But this DocHybridTime is
of no interest to us in key/value pairs of the regular RocksDB where
the commit time (from the key) is lower than the history cutoff used in
the compaction, so we can remove the intent hybrid time during the
compaction (read below for a longer explanation). This saves time
during the read path because we do not have to decode this intent
DocHybridTime in the value part of a RocksDB record. Read time for a
newly added test, PgMiniTest.BigReadWithCompaction, is 2.4s (without
this cleanup) and 2.3s (with this cleanup).

Here are some more details on why we need to store the DocHybridTime
(hybrid time + write id) of the original write intent that led to
creation of a RocksDB key/value pair in the regular RocksDB of the
tablet after the transaction was committed and its intents were
applied, and why it is no longer needed for records whose commit time
is lower than the history cutoff timestamp. A snapshot isolation read
operation running on behalf of a transaction on a particular tablet has
an associated pair of hybrid time values called global_limit (transaction
start time + max clock skew) and local_limit (the safe time observed by
the first operation of this transaction on this tablet, or the same as
global_limit if this is the first operation of this transaction on this
tablet). The read operation is trying to produce a view of data as of a
particular hybrid time read_time. For records in the regular RocksDB
that were written as a result of committing/applying another
transaction, we always include them in the result if the commit time
(stored as the DocHybridTime at the end of the RocksDB key) is <=
read_time. If it is greater than read_time, we need to decide whether
that other transaction could have been committed before our read
operation started. If there is a possibility of that, we have to
restart the read operation at a higher read time to include the effect
of this other transaction, which is detrimental to read performance.
Therefore, any way to prove that the other transaction must have
committed strictly after our read operation started is beneficial for
performance. If the other transaction's commit time (hybrid time from
the key of the regular RocksDB record) is greater than the read
operation's global_limit, then the other transaction certainly got
committed after our read operation started: the maximum hybrid time on
any server of the cluster at the moment our read operation started was
at most global_limit. However, another sufficient condition that allows
us to prove that we don't have to do a read restart is as follows.
If the intent of the other transaction that led to creation of the
current RocksDB key/value record in the regular RocksDB that we are
looking at was generated at a hybrid time higher than local_limit, i.e.
after the safe time on this tablet that we noted when an earlier RPC
request from the current transaction hit the tablet, then we know that
this other transaction could not have been committed before our read
operation started (i.e. it had not even finished writing intents by the
time our transaction already sent a read request to this tablet), and
we don't have to show this other transaction's effects in the read
operation's results, so no read restart is needed.

This is implemented as follows in `intent_aware_iterator.cc`. Note that
comparisons are being done on encoded DocHybridTime values (hybrid time
+ write id), so the timestamp order is the opposite of the byte
comparison order of encoded values. A key/value pair that we just found
where the intent hybrid time is below the local limit and the commit
hybrid time is below the global limit is considered valid, just as we
described above, and could lead to a read restart. If any of the two
conditions does not hold, we ignore the key/value pair and prevent the
possibility of a read restart.

  if (encoded_intent_doc_ht.compare(Slice(encoded_read_time_local_limit_)) > 0 &&
      encoded_doc_ht.compare(Slice(encoded_read_time_global_limit_)) > 0) {
     iter_valid_ = true;
     return;
  }

Now let's look at the interaction of this logic with compactions. After
we do a compaction with a particular history cutoff timestamp, we are
guaranteeing that no further snapshot isolation read operations are
allowed with read time lower than that history cutoff value. Therefore,
if we remove intent hybrid time from values where the commit hybrid
time in the key is lower than the history cutoff timestamp, we don't
lose any read restart prevention opportunities on the read path: those
opportunities would only exist if the read time turned out to be lower
than the commit timestamp of the transaction that produced this record.
But this record's commit timestamp is lower than the read time of any
future read operation, so it will always have to be included in the
read operation's results anyway.

Test Plan: ybd --gtest_filter PgMiniTest.BigReadWithCompaction

Reviewers: mikhail

Reviewed By: mikhail

Subscribers: kannan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8510