Fix default value of upgrade type in UI when doing software upgrade.
Summary:
D8443 which made changes to allow for in-memory gflag change inadvertantly caused the
software upgrade path's default option to be unset, causing the default state to be marked as false,
and causing software upgrades to by default be non-rolling.

Test Plan: Checked the UI and saw that the default was set to rolling.

Reviewers: ram, andrew, rahuldesirazu

Reviewed By: rahuldesirazu

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D8555